Fixed the KDTree 3-value constrctor to use all doubles
git-svn-id: https://java-algorithms-implementation.googlecode.com/svn/trunk@437 032fbc0f-8cab-eb90-e552-f08422b9a96a

+++ src/com/jwetherell/algorithms/data_structures/KdTree.java
@@ -556,7 +556,7 @@ public class KdTree<T extends KdTree.XYZPoint> {
             this.z = 0;
         }
 
-        public XYZPoint(double x, int y, double z) {
+        public XYZPoint(double x, double y, double z) {
             this.x = x;
             this.y = y;
             this.z = z;
